third_party/blink/web_tests/platform/mac-mac11/external/wpt/service-workers/service-worker/worker-interception-redirect.https-expected.txt

This is a testharness.js-based test.
[PASS] initialize global state
[FAIL] Case #1: network scope1->scope2 (classic DedicatedWorker)
  assert_unreached: Worker error Reached unreachable code
[FAIL] Case #1: network scope1->scope2 (module DedicatedWorker)
  assert_unreached: Worker error Reached unreachable code
[FAIL] Case #1: network scope1->scope2 (classic SharedWorker)
  assert_unreached: Worker error Reached unreachable code
[FAIL] Case #1: network scope1->scope2 (module SharedWorker)
  assert_unreached: Worker error Reached unreachable code
[FAIL] Case #2: network scope1->out-scope (classic DedicatedWorker)
  assert_unreached: Worker error Reached unreachable code
[FAIL] Case #2: network scope1->out-scope (module DedicatedWorker)
  assert_unreached: Worker error Reached unreachable code
[FAIL] Case #2: network scope1->out-scope (classic SharedWorker)
  assert_unreached: Worker error Reached unreachable code
[FAIL] Case #2: network scope1->out-scope (module SharedWorker)
  assert_unreached: Worker error Reached unreachable code
[FAIL] Case #3: sw scope1->scope2 (classic DedicatedWorker)
  assert_unreached: Worker error Reached unreachable code
[FAIL] Case #3: sw scope1->scope2 (module DedicatedWorker)
  assert_unreached: Worker error Reached unreachable code
[FAIL] Case #3: sw scope1->scope2 (classic SharedWorker)
  assert_unreached: Worker error Reached unreachable code
[FAIL] Case #3: sw scope1->scope2 (module SharedWorker)
  assert_unreached: Worker error Reached unreachable code
[FAIL] Case #4: sw scope1->out-scope (classic DedicatedWorker)
  assert_unreached: Worker error Reached unreachable code
[FAIL] Case #4: sw scope1->out-scope (module DedicatedWorker)
  assert_unreached: Worker error Reached unreachable code
[PASS] Case #4: sw scope1->out-scope (classic SharedWorker)
[PASS] Case #4: sw scope1->out-scope (module SharedWorker)
[PASS] cleanup global state
[FAIL] Case #4: sw scope1->out-scope (classic SharedWorker, importScripts())
  assert_equals: expected "sw1 saw importScripts from the worker: /service-workers/service-worker/resources/import-scripts-echo.py" but got "importScripts: served from network"
[FAIL] Case #4: sw scope1->out-scope (classic SharedWorker, fetch())
  assert_equals: expected "fetch(): sw1 saw the fetch from the worker: /service-workers/service-worker/resources/simple.txt" but got "fetch(): a simple text file\n"
[PASS] Case #4: sw scope1->out-scope (classic SharedWorker, location.href)
[PASS] Case #4: sw scope1->out-scope (module SharedWorker, importScripts())
[FAIL] Case #4: sw scope1->out-scope (module SharedWorker, fetch())
  assert_equals: expected "fetch(): sw1 saw the fetch from the worker: /service-workers/service-worker/resources/simple.txt" but got "fetch(): a simple text file\n"
[PASS] Case #4: sw scope1->out-scope (module SharedWorker, location.href)
Harness: the test ran to completion.
